" A review on Li-FI" in " International Journal on Emerging Trends in Electronics and Communication Engineering (Vol.2, Issue 1-2018)
This Paper introduces a new wireless communication method known as Li-Fi or Light Fidelity which was proposed by Harald Haas. It uses illumination for data transmission or light as a medium of communication. It uses normal LED lights having variation in its intensity which has speed, a human eye cannot follow. As this technology spectrum which comprises a wide range of frequencies, from the infrared through visible, down to ultraviolet spectrum for communication can produce a theoretical speed of 10 Gbps. So, it’s a much more better technology for data transmission compare existing Wi-Fi. Further research has to be done for better outcome.